--- cmd/ready/main.go
@@ -1,8 +1,0 @@
-package main
-
-import "os"
-
-func main() {
-	cli := &Ready{outStream: os.Stdout, errStream: os.Stderr}
-	os.Exit(cli.Run(os.Args))
-}
