+++ cmd/ready/makefile
@@ -1,0 +1,5 @@
+all: router
+
+.PHONY: router
+router:
+	go build -o router ready.go router.go
